Add files via upload
A teacher wants to buy supplies for her classroom. She needs 15 notebooks and 20 pens. Each notebook costs Rs. 2.25, and each pen costs Rs. 0.75. Write a C++ program that calculates the total cost of the supplies.

This was the program to help the teacher to find out the total price of pen and book with the help of c++ programming